% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledf.R
\docType{data}
\name{df}
\alias{df}
\title{Example Data Frame for Innovative Polygon Trend Analysis}
\format{
Data frame With Months in the first column, calculations from the first half values in the second column and calculations from the second half values in the third column.
}
\usage{
df
}
\description{
User is advised to prepare the data as suggested in the example to work with monthly data analysis using Innovative Polygon Trend Analysis Method.
}
\keyword{datasets}
