% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipta.R
\name{ipta}
\alias{ipta}
\title{Innovative Polygon Trend Analysis}
\usage{
ipta(x = x, y = y, Monthly = Monthly)
}
\arguments{
\item{x}{- First half of Time series data vector}

\item{y}{- First half of Time series data vector}

\item{Monthly}{- Months of Time series data vector}
}
\value{
Slope  - Slope of the trend

Distance  - Trend indicator
}
\description{
Innovative Polygon Trend Analysis method proposed by Sen et al. (2019) is implemented in this function.
This function is implemented to work with monthly time series. Users are advised to prepare input data frame as provided in example.
}
\examples{

ipta(x = df$First_half, y = df$Second_half, Monthly = df$Months)

}
\references{
Sen, Z., Sisman, E., Dabanli, I., 2019. Innovative Polygon Trend Analysis (IPTA) and applications. J. Hydrol. 575, 202-210. https://doi.org/10.1016/j.jhydrol.2019.05.028
}
