% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqmk.R
\name{sqmk}
\alias{sqmk}
\title{Sequential Mann-Kendall Plot}
\usage{
sqmk(x,startyear)
}
\arguments{
\item{x}{- Time series data vector}

\item{startyear}{- Starting of the time series (year!)}
}
\value{
Prograde  - Prograde series

Retrograde  - Retrograde series

Intersection  - True if intersecting and False if not intersecting.
}
\description{
The series is rearranged based on the ranks of the original series. Prograde and retrograde series plots are generated as per Sneyers (1990). The point where prograde and retrograde series intersect will approximately indicate the possible point of change.
}
\details{
The point where prograde series and retrograde series intersect, will indicate probable change point in the series. When there is no clear trend in the data, the series will intersect at several locations.
}
\examples{
x<-c(Nile)
sqmk(x,1871)

}
\references{
Sneyers S (1990) On the statistical analysis of series of observations. Technical note no 5 143, WMO No 725 415 Secretariat of the World Meteorological Organization, Geneva, 192 pp
}
